% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm_builder.R
\name{ecm_builder}
\alias{ecm_builder}
\title{Create simulations for long-term effects from error correction models (ECM)}
\usage{
ecm_builder(obj, baseline_df, lag_dv, lag_iv, d_iv, iv_shock, t_extent = 5,
  nsim = 1000, mu, Sigma)
}
\arguments{
\item{obj}{a fitted model object for an ECM model}

\item{baseline_df}{a data frame with fitted values for the baseline scenario
before the shock to \code{lag_dv}. Note column names should match coefficient
names in \code{obj}. Variables not include are assumed to be 0.
Note also that change variables should be 0.}

\item{lag_dv}{charachter string identifying the name of the lagged dependent
variable in \code{obj}.}

\item{lag_iv}{charachter string identifying the name of the lagged
independent variable in \code{obj} that experiences a shock.}

\item{d_iv}{charachter string identifying the name of the change in
\code{lag_dv} between the current and previous time, i.e. the size of the
one period shock to \code{lag_dv}.}

\item{iv_shock}{numeric shock to \code{iv}.}

\item{t_extent}{numeric specifying the time points from the shock to
simulate the long-term effects of the shock for. The default is 5 time
points.}

\item{mu}{an optional vector giving the means of the variables estimated
from an ECM model. If \code{obj} is supplied then \code{mu} is ignored.}

\item{Sigma}{an optional positive-definite symmetric matrix estimated
from an ECM model. The matrix specifies the covariance matrix of the
variables. If \code{obj}  is supplied then \code{Sigma} is ignored.}

\item{nsims}{numeric. Number of simulations to draw.}
}
\description{
Create simulations for long-term effects from error correction models (ECM)
}

