% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm_builder.R
\name{ecm_builder}
\alias{ecm_builder}
\title{Create simulations for long-term effects from error correction models (ECM)}
\usage{
ecm_builder(obj, baseline_df, lag_dv, lag_iv, d_iv, iv_shock, lag_iv_2, d_iv_2,
  iv_2_shock, lag_iv_interaction_term, d_iv_interaction_term, t_extent = 5,
  nsim = 1000, ci = 0.95, slim = TRUE, mu, Sigma)
}
\arguments{
\item{obj}{a fitted model object from an ECM model estimated with
\code{\link{lm}}.}

\item{baseline_df}{a data frame with fitted values for the baseline scenario
before the shock to \code{lag_dv}. Note column names should match coefficient
names in \code{obj}. Variables not include are assumed to be 0.
Note also that change variables should be 0.}

\item{lag_dv}{charachter string identifying the name of the lagged dependent
variable in \code{obj}.}

\item{lag_iv}{charachter string identifying the name of the lagged
independent variable in \code{obj} that experiences a shock.}

\item{d_iv}{charachter string identifying the name of the change in
\code{lag_dv} between the current and previous time, i.e. the size of the
one period shock to \code{lag_dv}.}

\item{iv_shock}{numeric shock to \code{iv}.}

\item{lag_iv_2}{character string identifying variable name of the second
lagged independent variable in an interaction with \code{lag_iv}, if
applicable.}

\item{d_iv_2}{character string identifying variable name of the second
shocked independent variable in an interaction with \code{lag_iv}, if
applicable.}

\item{iv_2_shock}{numeric shock to \code{iv_2}. If not specified, set to
0.}

\item{t_extent}{numeric specifying the time points from the shock to
simulate the long-term effects of the shock for. The default is 5 time
points.}

\item{nsim}{numeric. Number of simulations to draw.}

\item{ci}{the proportion of the central interval of the simulations to
return. Must be in (0, 1] or equivalently (0, 100]. Note: if \code{ci = 1}
then the full interval (i.e. 100 percent) is assumed.}

\item{slim}{logical indicating whether to (if \code{FALSE}) return all
simulations in the central interval specified by \code{ci} for each fitted
scenario or (if \code{TRUE}) just the minimum, median, and maxium values.
See \code{\link{qi_slimmer}} for more details.}

\item{mu}{an optional vector giving the means of the variables estimated
from an ECM model. If \code{obj} is supplied then \code{mu} is ignored.}

\item{Sigma}{an optional positive-definite symmetric matrix estimated
from an ECM model. The matrix specifies the covariance matrix of the
variables. If \code{obj}  is supplied then \code{Sigma} is ignored.
Note that if the model includes an intercept the corresponding column in
\code{Sigma} must be called \code{intercept_}.}
}
\description{
Create simulations for long-term effects from error correction models (ECM)
}

